move large function into separate file
